rename hooks to tags
